Add preference for enabling root access (3/3)
 * Root over ADB is now disabled by default on non-eng builds
 * Requires support in su binary and adbd
 * Root access can be configured to use ADB only, apps only, or all

Change-Id: I132a35a31c676511db954ff19120fa6d3c795b23

Settings: Make CM developer options additions respect the on/off toggle
This commit fixes the problem where, when dev options are turned off, our
CM customizations don't turn off with the rest of the options and the
dangerous option (root access) does not reset to a safe setting.

Change-Id: I879756ea353ce2940c05b224998fc1a945f2ceff

DevSettings: Fix the root-access toggle in eng builds

It was being unconditionally removed, put it back

Change-Id: I8eecab7353c59b224d917cc1710124f57656cf47

Hide Superuser entry if root access for apps is disabled.

Change-Id: Ibb77ba5a0fccf22564a26b8b1e13f7762c65718e