Settings: Don't hide home category if user has a settings capable launcher
By default the settings app hides the home category if it notices the user
has 1 launcher. Due to the ability to have settings per launcher, if the
user only has one launcher which happens to support settings, hiding this
category renders that feature useless.

So now we check if the user has < 2 launchers AND whether they support
a settings panel.

Change-Id: Id5dd94091a31fd9ee3637e0975816481903255bf