Settings : Don't disable persisting for MultiSelectListPreference
There is an aosp bug where MultiSelectListPreference stores a custom state type,
but doesn't override onRestoreState and causes an exception to be thrown when inside
activities that don't override orientation change.

This patch resolves this issue by persisting the preference change, which bypasses that route.
No functional impact since we override the values from the settings provider onresume anyway.

Repro :

    - Launch dialer
    - Launch Settings
    - Launch blacklist
    - Launch settings
    - Rotate phone
    - Realize you need this patch :)

Change-Id: Ie51119190fa6c43db612223bdb60c00fe607b9bf