Settings: clear dialog states when forgetting network
We need to clear out the saved dialog network states when forgetting
networks, otherwise on re-connect of the same network it can lead to a
state where user is presented with the incorrect (edit) dialog instead
of the new network connect dialog.

Repro steps:
1) begin connecting to any secured wifi network with incorrect password
2) while its "Authenticating" click the AP and click "Forget"
3) Immediately after returning to the list click on that same AP and see
   it opened with edit dialog

Even after editing the information (such as inputting password) it is
not saved and the user has to re-enter their information again after
hitting 'done'.

Ref: CYNGNOS-1209

Change-Id: I504a8a772a2e6cfa9d463dfbf41547bcf4fea5a6
Signed-off-by: Roman Birg <roman@cyngn.com>